Delay kernel fetch until advanced settings page
