+++ src/linodes/linode/settings/layouts/IndexPage.js
@@ -5,29 +5,10 @@ import { push } from 'react-router-redux';
 
 import { Tabs } from 'linode-components/tabs';
 
-import { setError } from '~/actions/errors';
-import { linodes } from '~/api';
-import { getObjectByLabelLazily } from '~/api/util';
-
 import { selectLinode } from '../../utilities';
 
 
 export class IndexPage extends Component {
-  static async preload({ dispatch, getState }, { linodeLabel }) {
-    if (window.location.pathname.indexOf('/settings/advanced') === -1) {
-      // No need to preload disks.
-      return;
-    }
-
-    const { id } = await dispatch(getObjectByLabelLazily('linodes', linodeLabel));
-
-    try {
-      await dispatch(linodes.disks.all([id]));
-    } catch (e) {
-      dispatch(setError(e));
-    }
-  }
-
   render() {
     const { linode } = this.props;
     if (!linode) return null;
